% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{append_permanent}
\alias{append_permanent}
\title{Append values to database}
\usage{
append_permanent(
  p,
  x,
  remote_name,
  data_type = c("matrix", "df", "polygon", "points"),
  fields = NULL,
  ...
)
}
\arguments{
\item{p}{Pool connection object}

\item{x}{data to append, given as a data.frame-like object}

\item{remote_name}{name to assign the object on the database}

\item{data_type}{type of data that is being appended to}

\item{fields}{additional fields constraints that might be desired if the table
is being newly generated}

\item{...}{additional params to pass to \code{\link[DBI]{dbCreateTable}}}
}
\description{
Writes values to database in an appending manner. As such, it does not check
if there is already an existing table.
}
