% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{as.matrix,dbMatrix-method}
\alias{as.matrix,dbMatrix-method}
\title{Convert dbMatrix to in-memory matrix}
\usage{
\S4method{as.matrix}{dbMatrix}(x, ...)
}
\arguments{
\item{x}{A dbMatrix object (dbSparseMatrix or dbDenseMatrix)}

\item{...}{Additional arguments (not used)}
}
\value{
A Matrix::dgCMatrix or base R matrix
}
\description{
Converts a dbMatrix object into an in-memory matrix or sparse matrix.
}
\details{
This method converts a dbMatrix object (dbSparseMatrix or dbDenseMatrix) to an in-memory
Matrix::dgCMatrix (for sparse) or base R matrix (for dense).
Warning: This can cause memory issues if the input matrix is large.
}
\concept{transform}
