% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.spatvector,dbSpatProxyData-method}
\alias{as.spatvector,dbSpatProxyData-method}
\title{Convert to terra spatvector}
\usage{
\S4method{as.spatvector}{dbSpatProxyData}(x, ...)
}
\description{
Convert to terra spatvector
}
