% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{backendID-generic}
\alias{backendID-generic}
\alias{backendID,}
\alias{backendID<-}
\alias{backendID,dbData-method}
\alias{backendID<-,backendInfo,character-method}
\alias{backendID<-,dbData,character-method}
\title{Get and set hash ID of backend}
\usage{
\S4method{backendID}{backendInfo}(x)

\S4method{backendID}{dbData}(x)

\S4method{backendID}{backendInfo,character}(x, ...) <- value

\S4method{backendID}{dbData,character}(x, ...) <- value
}
\arguments{
\item{x}{object to get hash from}

\item{value}{(character) hash ID to set}
}
\description{
Get the hash value/slot
}
