% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{backendInfo}
\alias{backendInfo}
\title{backendInfo}
\description{
Simple S4 class to contain information about the database backend and
regenerate connection pools. A hash is generated from the db_path to act
as a unique identifier for each backend.
}
\section{Slots}{

\describe{
\item{\code{driver_call}}{DB driver call stored as a string}

\item{\code{db_path}}{path to database}

\item{\code{add_params}}{additional connection params e.g. user, pass, etc}

\item{\code{hash}}{xxhash64 hash of the db_path}
}}

