% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{backendSize}
\alias{backendSize}
\title{Size of backend database}
\usage{
backendSize(backend_ID)
}
\arguments{
\item{backend_ID}{backend ID}
}
\description{
Given a backend ID, find the current size of the DB backend file
}
