% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{callback_combineCols}
\alias{callback_combineCols}
\title{Combine columns values}
\usage{
callback_combineCols(x, col_indices, new_col = "new", remove_originals = TRUE)
}
\arguments{
\item{x}{data.table}

\item{col_indices}{numeric vector. Col indices to combine}

\item{new_col}{name of new combined column}

\item{remove_originals}{remove originals cols of the combined col
default = TRUE}
}
\value{
data.table with specified column combined
}
\description{
Combine columns values
}
