% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{callback_formatIJX}
\alias{callback_formatIJX}
\title{Convert table to ijx}
\usage{
callback_formatIJX(x, group_by = 1)
}
\arguments{
\item{x}{data.table}

\item{group_by}{numeric or character designating which column to set as i.
Default = 1st column}
}
\value{
data.table in ijx format
}
\description{
Building block function intended for use as or in callback functions used
when streaming large flat files to database. Converts the input data.table
to long format with columns i, j, and x where i and j are row and col
names and x is values. Columns i and j are additionally set as 'character'.
}
