% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{callback_swapCols}
\alias{callback_swapCols}
\title{Swap values in two columns}
\usage{
callback_swapCols(x, c1, c2)
}
\arguments{
\item{x}{data.table}

\item{c1}{col 1 to use (character)}

\item{c2}{col 2 to use (character)}
}
\value{
data.table with designated column values swapped
}
\description{
Swap values in two columns
}
