% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{castNumeric}
\alias{castNumeric}
\alias{castNumeric,dbMatrix,missing-method}
\title{Set a column to numeric}
\usage{
\S4method{castNumeric}{dbData,character}(x, col, ...)

\S4method{castNumeric}{dbMatrix,missing}(x, col, ...)
}
\arguments{
\item{x}{Duckling data object}

\item{col}{column to cast to numeric}

\item{...}{additional params to pass}
}
\description{
Sets a column to numeric after first checking the column data type. Does
nothing if the column is already a \code{double}
This precaution is to avoid truncation of values.
}
