% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_chunking.R
\name{chunk_plan}
\alias{chunk_plan}
\title{Plan spatial chunking extents}
\usage{
chunk_plan(extent, min_chunks = NULL, nrows = NULL, ncols = NULL)
}
\arguments{
\item{extent}{terra SpatExtent that covers the region to spatially chunk}

\item{min_chunks}{numeric. minimum number of chunks to use.}

\item{nrows, ncols}{numeric. nrow/ncol must be provided as a pair. Determines how many
rows and cols respectively will be used in spatial chunking. If NULL, min_chunks
will be used as an automated method of planning the spatial chunking}
}
\value{
vector of chunked SpatExtents
}
\description{
Generate the individual extents that will be used to spatially chunk a set of
data for piecewise and potentially parallelized processing. Chunks will be
generated first by row, then by column. The chunks try to be as square as
possible since downstream functions may require slight expansions of the
extents to capture all parts of selected polygons. Minimizing the perimeter
relative to area decreases waste.
}
\examples{
\dontrun{
 e <- terra::ext(0, 100, 0, 100)

e_chunk1 <- chunk_plan(e, min_chunks = 9)
e_poly1 <- sapply(e_chunk1, terra::as.polygons)
e_poly1 <- do.call(rbind, e_poly1)
plot(e_poly1)

e_chunk2 <- chunk_plan(e, nrows = 3, ncols = 5)
e_poly2 <- sapply(e_chunk2, terra::as.polygons)
e_poly2 <- do.call(rbind, e_poly2)
plot(e_poly2)
}
}
\seealso{
\code{\link{get_dim_n_chunks}}
}
\keyword{internal}
