% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_chunking.R
\name{chunk_spat_apply}
\alias{chunk_spat_apply}
\title{chunk_plan slightly expands bounds, allowing for use of 'soft' selections
with 'extent_filter() on two sides during the chunk processing
Calculations with y are expected to be performed relative to x. Spatial
chunk subsetting of y is performed based on the updated extent of the x
chunks after their chunk subset.
1. Setup lapply
2. Run provided function
3. write or return values}
\usage{
chunk_spat_apply(
  x = NULL,
  y = NULL,
  chunk_y = TRUE,
  fun,
  extent = NULL,
  n_per_chunk = 1e+05,
  remote_name = NULL,
  output = c("tbl", "dbPolygonProxy", "dbPointsProxy"),
  progress = TRUE,
  ...
)
}
\description{
chunk_plan slightly expands bounds, allowing for use of 'soft' selections
with 'extent_filter() on two sides during the chunk processing
Calculations with y are expected to be performed relative to x. Spatial
chunk subsetting of y is performed based on the updated extent of the x
chunks after their chunk subset.
1. Setup lapply
2. Run provided function
3. write or return values
}
