% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{closeBackend}
\alias{closeBackend}
\title{Close backend connection pool}
\usage{
closeBackend(hash)
}
\arguments{
\item{hash}{hash of backend to close (optional)}
}
\description{
Closes pools. If specific hash(es) are given then those will be closed. When
no specific hash is given, all existing backends will be closed.
}
