% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{colTypes,dbMatrix-method}
\alias{colTypes,dbMatrix-method}
\title{Return the column types of a `dbMatrix` object}
\usage{
\S4method{colTypes}{dbMatrix}(x, ...)
}
\description{
Return the column types of a `dbMatrix` object
}
\concept{matrix_props}
