% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckDB_backend.R
\name{computeDBMatrix}
\alias{computeDBMatrix}
\title{Execute dplyr query of dbMatrix and save the results in remote database}
\usage{
computeDBMatrix(
  x,
  remote_name = "test",
  temporary = TRUE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{dbData object to compute from}

\item{remote_name}{name of table to create on DB}

\item{temporary}{(default = TRUE) whether to make a temporary table on the DB}

\item{overwrite}{(default = FALSE) whether to overwrite if remote_name already exists}

\item{...}{additional params to pass}
}
\description{
Calculate the lazy query of a dbMatrix object and send it to the database backend
either as a temporary or permanent table.
}
