% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckDB_backend.R
\name{con-generic}
\alias{con-generic}
\alias{con,dbMatrix-method}
\alias{con,dbDataFrame-method}
\title{Database connection}
\usage{
\S4method{con}{dbMatrix}(x)

\S4method{con}{dbDataFrame}(x)
}
\description{
Return the database connection object
}
