% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connection-generic}
\alias{connection-generic}
\alias{connection,}
\alias{connection<-}
\alias{connection<-,dbData-method}
\title{Database connection}
\usage{
\S4method{connection}{dbData}(x)

\S4method{connection}{dbData}(x) <- value
}
\arguments{
\item{x}{GiottoDB object}
}
\description{
Return the database connection object
}
