% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{create_dbSparseMatrix}
\alias{create_dbSparseMatrix}
\title{Create a dbSparseMatrix object from a sparse matrix}
\usage{
create_dbSparseMatrix(sparse_mat, con)
}
\arguments{
\item{sparse_mat}{A sparse matrix object}

\item{con}{A database connection object}
}
\value{
A dbSparseMatrix object
}
\description{
Internal function to create a dbSparseMatrix object
}
