% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{createBackend}
\alias{createBackend}
\title{Create GiottoDB backend}
\usage{
createBackend(
  drv = duckdb::duckdb(),
  dbdir = ":temp:",
  extension = ".duckdb",
  with_login = FALSE,
  ...
)
}
\arguments{
\item{drv}{database driver (default is duckdb::duckdb())}

\item{dbdir}{directory to create a backend database}

\item{extension}{file extension (default = '.duckdb')}

\item{with_login}{(default = FALSE) whether a login is needed}

\item{...}{additional params to pass to pool::dbPool()}
}
\value{
environment containing backendInfo object
}
\description{
Defines and creates a database connection pool to be used by the backend.
This pool object and a backendInfo object that contains details about the
}
