% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{createDBMatrix}
\alias{createDBMatrix}
\title{Create a sparse or dense dbMatrix object}
\usage{
createDBMatrix(
  value,
  class = NULL,
  db_path = ":memory:",
  overwrite = FALSE,
  name = "dbMatrix",
  dims = NULL,
  dim_names = NULL,
  ...
)
}
\arguments{
\item{value}{data to be added to the database. See details for supported data types \code{(required)}}

\item{class}{class of the dbMatrix: \code{dbDenseMatrix} or \code{dbSparseMatrix} \code{(required)}}

\item{db_path}{path to database on disk (relative or absolute) or in memory \code{(":memory:")}}

\item{overwrite}{whether to overwrite if table already exists in database \code{(required)}}

\item{name}{table name to assign within database \code{(optional, default: "dbMatrix")}}

\item{dims}{dimensions of the matrix \code{(optional: [int, int])}}

\item{dim_names}{dimension names of the matrix \code{(optional: list(enum, enum))}}

\item{...}{additional params to pass}
}
\description{
Create an S4 \code{dbMatrix} object in sparse or dense triplet vector format.
}
\details{
This function reads in data into a pre-existing DuckDB database. Based
on the \code{name} and \code{db_path} a lazy connection is then made
downstream during \code{dbMatrix} initialization.

Supported \code{value} data types:
\itemize{
 \item \code{dgCMatrix} In-memory sparse matrix from the \code{Matrix} package
 \item \code{matrix} In-memory dense matrix from base R
 \item \code{.mtx} Path to .mtx file (TODO)
 \item \code{.csv} Path to .csv file (TODO)
 \item \code{tbl_duckdb_connection} Table in DuckDB database in ijx format from
 existing \code{dbMatrix} object. \code{dims} and \code{dim_names} must be
 specified if \code{value} is \code{tbl_duckdb_connection}.
}
}
\examples{
dgc <- dbMatrix:::sim_dgc()
dbSparse <- createDBMatrix(value = dgc, db_path = ":memory:",
                           name = "sparse_matrix", class = "dbSparseMatrix",
                           overwrite = TRUE)
dbSparse
}
