% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatProxy.R
\name{createDBPolygonProxy}
\alias{createDBPolygonProxy}
\title{Create a framework for chunked processing with terra polygon SpatVectors}
\usage{
createDBPolygonProxy(
  SpatVector,
  remote_name = "poly_test",
  db_path = ":temp:",
  attributes,
  extent,
  n_poly,
  ...
)
}
\arguments{
\item{SpatVector}{object coercible to SpatVector or filepath to spatial data
readable by \code{\link[terra]{vect}}}
}
\description{
Create an S4 dbPolygonProxy object that is composed of two database tables.
One table hold geometry information while the other holds attribute information.
}
\details{
Information is only read into the database during this process. Based
on the \code{remote_name} and \code{db_path} a lazy connection is then made
downstream during dbData intialization and appended to the object. If the data
already exists within the database backend then it is entirely permissible to
omit the \code{SpatVector} param.
}
