% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbDataFrame-class}
\alias{dbDataFrame-class}
\alias{dbDataFrame}
\title{S4 dbDataFrame class}
\description{
Representation of dataframes using an on-disk database. Each object
is used as a connection to a single table that exists within the database.
}
\section{Slots}{

\describe{
\item{\code{driver}}{driver function for the database}

\item{\code{connection}}{database connection object}

\item{\code{path}}{path to database on-disk file}

\item{\code{table_name}}{name of table within database that contains matrix}

\item{\code{valid}}{whether the object is initialized for use}
}}

