% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbDisconnect,dbMatrix-method}
\alias{dbDisconnect,dbMatrix-method}
\alias{dbListTables,dbMatrix-method}
\title{dbDisconnect}
\usage{
\S4method{dbDisconnect}{dbMatrix}(x, ...)

\S4method{dbListTables}{dbMatrix}(x, ...)
}
\description{
dbDisconnect

dbListTables
}
