% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbList,DBIConnection-method}
\alias{dbList,DBIConnection-method}
\title{List remote tables, temporary tables, and views}
\usage{
\S4method{dbList}{DBIConnection}(conn)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}
}
\description{
Pretty prints tables, temporary tables, and views in the database.
}
\details{
Similar to DBI::dbListTables, but categorizes tables into three categories:
\itemize{
\item Tables
\item Temporary Tables (these will be removed when the connection is closed)
\item Views (these may be removed when the connection is closed)
}
}
\concept{dbData}
