% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{dbMatrix_from_tbl}
\alias{dbMatrix_from_tbl}
\title{dbMatrix_from_tbl}
\usage{
dbMatrix_from_tbl(
  tbl,
  rownames_colName,
  colnames_colName,
  name = "dbMatrix",
  overwrite = FALSE
)
}
\arguments{
\item{tbl}{\code{tbl_duckdb_connection} table in DuckDB database in long format}

\item{rownames_colName}{\code{character} column name of rownames in tbl \code{(required)}}

\item{colnames_colName}{\code{character} column name of colnames in tbl \code{(required)}}

\item{name}{table name to assign within database \code{(required, default: "dbMatrix")}}

\item{overwrite}{whether to overwrite if table already exists in database \code{(required)}}

\item{con}{DBI or duckdb connection object \code{(required)}}
}
\value{
\code{dbMatrix} object
}
\description{
Construcst a \code{dbSparseMatrix} object from a \code{tbl_duckdb_connection} object.
}
\details{
The \code{tbl_duckdb_connection} object must contain dimension names.
}
\keyword{internal}
