% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbMemoryLimit}
\alias{dbMemoryLimit}
\title{Get and set DB memory limits}
\usage{
dbMemoryLimit(x, limit, ...)
}
\arguments{
\item{x}{backend_ID or pool object}

\item{limit}{character. Memory limit to use with units included (for example
'10GB'). If missing, will get the current setting. If 'RESET' will reset to
default}

\item{...}{additional params to pass}
}
\description{
Get and set DB memory limits
}
