% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dbms-generic}
\alias{dbms-generic}
\alias{dbms}
\alias{dbms,character-method}
\alias{dbms,Pool-method}
\alias{dbms,DBIConnection-method}
\title{Database management system}
\usage{
\S4method{dbms}{dbData}(x, ...)

\S4method{dbms}{character}(x, ...)

\S4method{dbms}{Pool}(x, ...)

\S4method{dbms}{DBIConnection}(x, ...)
}
\arguments{
\item{x}{A connection pool object, DBI connection, or dbData}

\item{...}{additional params to pass}
}
\description{
Get the database management system (dbms) of an object
}
