% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbPointsProxy-class}
\alias{dbPointsProxy-class}
\alias{dbPointsProxy}
\title{S4 dbPointsProxy class}
\description{
Representation of point information using an on-disk database. Intended to
be used to store information that can be pulled into terra point SpatVectors
}
\section{Slots}{

\describe{
\item{\code{attributes}}{dbDataFrame of attributes information}

\item{\code{n_points}}{number of points}

\item{\code{feat_ID}}{feature IDs}

\item{\code{extent}}{extent of points}
}}

