% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbPolygonProxy-class}
\alias{dbPolygonProxy-class}
\alias{dbPolygonProxy}
\title{S4 dbPolygonProxy class}
\description{
Representation of polygon information using an on-disk database. Intended to
be used to store information that can be pulled into terra polygon SpatVectors
}
\section{Slots}{

\describe{
\item{\code{attributes}}{dbDataFrame of attributes information}

\item{\code{n_poly}}{number of polygons}

\item{\code{extent}}{extent of polygons}
}}

