% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbSettings}
\alias{dbSettings}
\alias{dbSettings,ANY,character,ANY-method}
\title{Get and set database settings}
\usage{
\S4method{dbSettings}{ANY,character,missing}(x, setting, value, ...)

\S4method{dbSettings}{ANY,character,ANY}(x, setting, value, ...)
}
\arguments{
\item{x}{backend_ID or pool object}

\item{setting}{character. Setting to get or set}

\item{value}{if missing, will retrieve the setting. If provided, will attempt
to set the new value. If 'RESET' will reset the value to default}

\item{...}{additional params to pass}
}
\description{
Get and set database settings
}
