% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbSparseMatrix-class}
\alias{dbSparseMatrix-class}
\alias{dbSparseMatrix}
\title{S4 Class for dbSparseMatrix}
\description{
Representation of sparse matrices using an on-disk database.
Inherits from dbMatrix.
}
\section{Slots}{

\describe{
\item{\code{data}}{An ijx matrix without zeros}
}}

