% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbThreads}
\alias{dbThreads}
\title{Set and get number of threads to use in database backend}
\usage{
dbThreads(x, threads, ...)
}
\arguments{
\item{x}{backend ID or pool object of backend}

\item{threads}{numeric or integer. Number of threads to use.
If missing, will get the current setting. If 'RESET' will reset to default.}

\item{...}{additional params to pass}
}
\description{
Set and get number of threads to use in database backend
}
