% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckDB_backend.R
\name{deleteTableRow}
\alias{deleteTableRow}
\title{Delete row from table}
\usage{
deleteTableRow(x, y, verbose = TRUE)
}
\arguments{
\item{x}{dbData object}

\item{y}{data.table of values (by column) to remove from x by matching}

\item{verbose}{be verbose}
}
\description{
Delete row from table
}
