% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{desparse_to_grid_bin_data}
\alias{desparse_to_grid_bin_data}
\title{Desparsify binned data and assign to rasterized locations}
\usage{
desparse_to_grid_bin_data(data, px_x, px_y)
}
\arguments{
\item{data}{data.frame}

\item{extent}{terra extent in which to plot}

\item{resolution}{min number of tiles to collect}
}
\description{
Desparsify binned data and assign to rasterized locations
}
\keyword{internal}
