% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from database}
\usage{
\S4method{disconnect}{dbData}(x, shutdown = TRUE)
}
\arguments{
\item{x}{Duckling object}
}
\description{
Disconnects from a database. Closes both the connection and driver objects
by default
}
