% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.check_con}
\alias{.check_con}
\title{Input validation for conn arg}
\usage{
.check_con(conn)
}
\arguments{
\item{conn}{A \link{tbl_duckdb_connection} object}
}
\description{
Input validation for conn arg
}
\keyword{internal}
