% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.check_overwrite}
\alias{.check_overwrite}
\title{Input validation for overwrite arg}
\usage{
.check_overwrite(conn, overwrite, name, skip_value_check)
}
\description{
Input validation for overwrite arg
}
\keyword{internal}
