% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.check_value}
\alias{.check_value}
\title{Input validation for data arg}
\usage{
.check_value(value)
}
\arguments{
\item{x}{A \link{Matrix}, \link{matrix}, or \link{tbl_duckdb_connection} object}
}
\description{
Input validation for data arg
}
\keyword{internal}
