% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{driver-generic}
\alias{driver-generic}
\alias{driver,}
\alias{driver<-}
\alias{driver<-,dbData-method}
\title{Database driver generation}
\usage{
\S4method{driver}{dbData}(x)

\S4method{driver}{dbData}(x) <- value
}
\description{
Set or get the driver/database instance generation function call
}
