% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{dropTableBE}
\alias{dropTableBE}
\title{Drop a table from the database}
\usage{
dropTableBE(conn, remote_name)
}
\arguments{
\item{conn}{connection object or pool}

\item{remote_name}{name of table to drop}
}
\description{
Drop a table from the database
}
