% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatProxy.R
\name{extent_calculate}
\alias{extent_calculate}
\title{Calculate dbSpatProxyData spatial extent}
\usage{
\S4method{extent_calculate}{dbSpatProxyData}(x, ...)
}
\arguments{
\item{x}{a dbSpatProxyData object}
}
\value{
terra SpatExtent
}
\description{
Calculate dbSpatProxyData spatial extent
}
\seealso{
Other Extent processing functions: 
\code{\link{extent_filter}()}
}
\concept{Extent processing functions}
\keyword{internal}
