% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_preset.R
\docType{data}
\name{fields_preset}
\alias{fields_preset}
\title{Table fields presets}
\format{
An object of class \code{list} of length 2.
}
\usage{
fields_preset
}
\description{
Fields presets to use with DBI::dbCreateTable() to set up remote tables and
their constraints
}
\details{
Package presets for DB table fields that can be easily accessed
}
\keyword{datasets}
