% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{flex_window_order}
\alias{flex_window_order}
\title{flex_window_order}
\usage{
flex_window_order(x, order_cols)
}
\arguments{
\item{x}{dbplyr::tbl_lazy}

\item{order_cols}{character vector of column names}
}
\description{
workaround for multiple dbplyr window column ordering
}
\keyword{internal}
