% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{fstreamToDB_IJX}
\alias{fstreamToDB_IJX}
\title{Stream large flat filess to database backend using fread}
\usage{
fstreamToDB_IJX(
  path,
  backend_ID,
  remote_name = "test",
  nlines = 10000L,
  cores = 1L,
  callback = NULL,
  overwrite = FALSE,
  with_pk = FALSE,
  ...
)
}
\arguments{
\item{path}{path to the matrix file}

\item{backend_ID}{ID of the backend to use}

\item{remote_name}{name to assign table of read in values in database}

\item{nlines}{integer. Number of lines to read per chunk}

\item{cores}{fread cores to use}

\item{callback}{callback functions to apply to each data chunk before it is
sent to the database backend}

\item{overwrite}{whether to overwrite if table already exists (default = FALSE)}

\item{with_pk}{whether to generate a primary key on i and j. Warning size will
increase greatly with a pk}

\item{...}{additional params to pass}
}
\description{
Stream large flat filess to database backend using fread
}
