% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_dbdir}
\alias{get_dbdir}
\title{get_dbdir}
\usage{
get_dbdir(dbData)
}
\arguments{
\item{dbData}{}
}
\description{
get_dbdir
}
\keyword{database}
\keyword{directory}
\keyword{internal}
\keyword{of}
\keyword{path}
\keyword{string}
\keyword{the}
\keyword{to}
