% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_chunking.R
\name{get_dim_n_chunks}
\alias{get_dim_n_chunks}
\title{Get rows and cols needed to create at least n chunks from given extent}
\usage{
get_dim_n_chunks(n, e)
}
\arguments{
\item{n}{minimum n chunks}

\item{e}{selection extent}
}
\value{
numeric vector of x and y stops needed
}
\description{
Algorithm to determine how to divide up a provided extent into
at least \code{n} different chunks. The chunks are arranged so as to prefer
being as square as posssible with the provided dimensions and minimum n chunks.
}
\examples{
\dontrun{
e <- terra::ext(0, 100, 0, 100)
get_dim_n_chunk(n = 5, e = e)
}
}
\seealso{
\code{\link{chunk_plan}}
}
