% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckDB_backend.R
\name{get_full_table_name_quoted}
\alias{get_full_table_name_quoted}
\title{Get the full table name consisting of the schema and table name}
\usage{
get_full_table_name_quoted(conn, remote_name)
}
\arguments{
\item{conn}{a hashID, DBI connection object, or pool object}

\item{remote_name}{name of table within DB}
}
\value{
return the full table name
}
\description{
Get the full name of a table. This information is then quoted to protect against
SQL injections.
Either param x or hash must be given
}
\keyword{internal}
