% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{get_MM_dim}
\alias{get_MM_dim}
\title{get_MM_dim}
\usage{
get_MM_dim(mtx_file_path)
}
\arguments{
\item{mtx_file_path}{path to .mtx file to be read into database}
}
\value{
integer vector of dimensions
}
\description{
Internal function to read dimensions of a .mtx file
}
\details{
Scans for the header of an mtx file (starting with %) and takes one more line
representing the dimensions and number of nonzero values.

Note: the header size can vary depending on the .mtx file.
}
\keyword{internal}
