% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{getBackendPool}
\alias{getBackendPool}
\alias{getBackendInfo}
\alias{getBackendConn}
\alias{getBackendPath}
\title{Get backend connection pool}
\usage{
getBackendPool(backend_ID)

getBackendInfo(backend_ID)

getBackendConn(backend_ID)

getBackendPath(backend_ID)
}
\arguments{
\item{backend_ID}{hashID generated from the dbpath (xxhash64) used as unique ID for
the backend}
}
\description{
Get backend information. \code{getBackendPool} gets the associated
connection pool object. \code{getBackendInfo} gets the backendInfo object
that contains all connection details needed to regenerate another connection
}
\section{Functions}{
\itemize{
\item \code{getBackendInfo()}: get backendInfo object containing connection details

\item \code{getBackendConn()}: Get a DBI connection object from pool.
Provided for convenience. Must be returned using pool::poolReturn() after use.

\item \code{getBackendPath()}: get filepath of backend

}}
