% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDBPath}
\alias{getDBPath}
\title{Get the database backend path}
\usage{
getDBPath(path = ":temp:", extension = ".duckdb")
}
\arguments{
\item{path}{directory path in which to place the backend}

\item{extension}{file extension of the backend (default is .duckdb)}
}
\description{
Get the full normalized filepath of a Giotto backend database. Additionally
passing path = ':memory:' will directly return ':memory:' and ':temp:' will
will have the function check tempdir() for the backend.
}
