% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_giotto.R
\name{giotto_libNorm_dbMatrix}
\alias{giotto_libNorm_dbMatrix}
\title{Giotto dbMatrix library normalization}
\usage{
giotto_libNorm_dbMatrix(dbMatrix, scalefactor)
}
\arguments{
\item{dbMatrix}{dbMatrix to use}

\item{scalefactor}{scalefactor to use}
}
\description{
Workaround for SQL faster PARTITION aggregation via a
GROUP BY and INNER JOIN workflow for dbMatrix. Compatible with dplyr chains
}
