% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R, R/names.R, R/operations.R
\name{[,dbMatrix,missing,missing,missing-method}
\alias{[,dbMatrix,missing,missing,missing-method}
\alias{[<-,dbMatrix,missing,missing,ANY-method}
\alias{[,dbMatrix,dbIndex,missing,missing-method}
\alias{[,dbMatrix,missing,dbIndex,missing-method}
\alias{[,dbMatrix,dbIndex,dbIndex,missing-method}
\alias{rownames,dbMatrix-method}
\alias{rownames<-,dbMatrix-method}
\alias{colnames,dbMatrix-method}
\alias{colnames<-,dbMatrix,ANY-method}
\alias{dimnames,dbMatrix-method}
\alias{dimnames<-,dbMatrix,list-method}
\alias{Arith,dbMatrix,ANY-method}
\alias{Arith,ANY,dbMatrix-method}
\alias{Arith,dbMatrix,dbMatrix-method}
\alias{Ops,dbMatrix,ANY-method}
\alias{Ops,ANY,dbMatrix-method}
\alias{Ops,dbMatrix,dbMatrix-method}
\alias{rowSums,dbDenseMatrix-method}
\alias{rowSums,dbSparseMatrix-method}
\alias{colSums,dbDenseMatrix-method}
\alias{colSums,dbSparseMatrix-method}
\alias{rowMeans,dbDenseMatrix-method}
\alias{rowMeans,dbSparseMatrix-method}
\alias{colMeans,dbDenseMatrix-method}
\alias{colMeans,dbSparseMatrix-method}
\alias{colSds,dbDenseMatrix-method}
\alias{colSds,dbSparseMatrix-method}
\alias{rowSds,dbDenseMatrix-method}
\alias{rowSds,dbSparseMatrix-method}
\alias{mean,dbDenseMatrix-method}
\alias{mean,dbSparseMatrix-method}
\alias{log,dbMatrix-method}
\alias{t,dbMatrix-method}
\alias{ncol,dbMatrix-method}
\alias{dim,dbMatrix-method}
\title{rowSums}
\usage{
\S4method{[}{dbMatrix,missing,missing,missing}(x, i, j)

\S4method{[}{dbMatrix,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{dbMatrix,dbIndex,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{dbMatrix,missing,dbIndex,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{dbMatrix,dbIndex,dbIndex,missing}(x, i, j, ..., drop = TRUE)

\S4method{rownames}{dbMatrix}(x)

\S4method{rownames}{dbMatrix}(x)

\S4method{rownames}{dbMatrix}(x) <- value

\S4method{colnames}{dbMatrix}(x)

\S4method{colnames}{dbMatrix}(x)

\S4method{colnames}{dbMatrix,ANY}(x) <- value

\S4method{dimnames}{dbMatrix}(x)

\S4method{dimnames}{dbMatrix,list}(x) <- value

\S4method{Arith}{dbMatrix,ANY}(e1, e2)

\S4method{Arith}{ANY,dbMatrix}(e1, e2)

\S4method{Arith}{dbMatrix,dbMatrix}(e1, e2)

\S4method{Ops}{dbMatrix,ANY}(e1, e2)

\S4method{Ops}{ANY,dbMatrix}(e1, e2)

\S4method{Ops}{dbMatrix,dbMatrix}(e1, e2)

\S4method{rowSums}{dbDenseMatrix}(x, ...)

\S4method{rowSums}{dbSparseMatrix}(x, ...)

\S4method{colSums}{dbDenseMatrix}(x, ...)

\S4method{colSums}{dbSparseMatrix}(x, ...)

\S4method{rowMeans}{dbDenseMatrix}(x, ...)

\S4method{rowMeans}{dbSparseMatrix}(x, ...)

\S4method{colMeans}{dbDenseMatrix}(x, ...)

\S4method{colMeans}{dbSparseMatrix}(x, ...)

\S4method{colSds}{dbDenseMatrix}(x, ...)

\S4method{colSds}{dbSparseMatrix}(x, ...)

\S4method{rowSds}{dbDenseMatrix}(x, ...)

\S4method{rowSds}{dbSparseMatrix}(x, ...)

\S4method{mean}{dbDenseMatrix}(x, ...)

\S4method{mean}{dbSparseMatrix}(x, ...)

\S4method{log}{dbMatrix}(x, ...)

\S4method{t}{dbMatrix}(x)

\S4method{ncol}{dbMatrix}(x)

\S4method{dim}{dbMatrix}(x)
}
\description{
rowSums

rowSums

colSums

colSums

rowMeans

rowMeans

colMeans

colMeans

colSds

colSds

rowSds

rowSds

mean

mean

log

Transpose

ncol

dim
}
