% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{is_init-generic}
\alias{is_init-generic}
\title{Determine if dbData object is initialized}
\usage{
\S4method{is_init}{dbData}(x, ...)
}
\arguments{
\item{x}{dbData object}
}
\description{
Determine if dbData object is initialized
}
\keyword{internal}
