% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDataFrame.R
\name{keyCol}
\alias{keyCol}
\alias{keyCol<-,dbDataFrame,character-method}
\title{Get and set a key column}
\usage{
\S4method{keyCol}{dbDataFrame}(x, ...)

\S4method{keyCol}{dbDataFrame,character}(x, ...) <- value
}
\arguments{
\item{x}{dbDataFrame}

\item{value}{character. Column to set as key}
}
\description{
Set a column as key in order to sort and subset by rows with dbDataFrame.
More than one column can be set at the same time, which will sort the table
by multiple columns (ascending), starting with the first specified.
}
\examples{
d <- simulate_dbDataFrame()
keyCol(d) <- c('Sepal.Length', 'Sepal.Width')
d[1:6,]
}
