% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{make_ijx_dimnames}
\alias{make_ijx_dimnames}
\title{Create database table with ijx and dimnames}
\usage{
make_ijx_dimnames(dbMatrix, name, overwrite = FALSE, colName_i, colName_j)
}
\arguments{
\item{dbMatrix}{dbMatrix object}

\item{name}{name of table to add to database}

\item{overwrite}{whether to overwrite if table already exists in database
default: 'FALSE'.'}

\item{colName_i}{name of column rownames to add to database}

\item{colName_j}{name of column colnames to add to database}
}
\description{
Create database table with ijx and dimnames
}
\details{
Constructs a table in a database that contains the accompanying dimnames
for a dbMatrix. The resulting columns in the table:
* i (row index)
* colName_i (rownames),
* j (col index)
* j_names (colnames)
* x (counts of i,j occcurences)
}
\keyword{internal}
