% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{nrow,dbMatrix-method}
\alias{nrow,dbMatrix-method}
\alias{ncol,dbMatrix-method}
\title{The Number of Rows/Columns of an Array}
\usage{
\S4method{nrow}{dbMatrix}(x)

\S4method{ncol}{dbMatrix}(x)
}
\description{
See ?\link{\code{base::nrow}} for more details.

See ?\link{\code{base::ncol}} for more details.
}
\concept{matrix_props}
