% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute.R
\name{precompute}
\alias{precompute}
\title{Precompute a dbMatrix table in a database}
\usage{
precompute(conn, m, n, name, verbose = TRUE)
}
\arguments{
\item{conn}{duckdb database connection}

\item{m}{number of rows of precomputed dbMatrix table}

\item{n}{number of columns of precomputed dbMatrix table}

\item{name}{name of the precomputed dbMatrix table to be created}

\item{verbose}{logical, print progress messages. default: TRUE.}
}
\value{
tbl_dbi
}
\description{
Precomputes a dbMatrix table in a specificied database connection.
This can speed up operations that involve breaking
sparsity of a \code{dbSparseMatrix},
such as in cases when performing + or - arithmetic operations.
Need only be run once.
}
\details{
The \code{num_rows} and \code{num_cols} parameters must exceed the
maximum row and column indices of the \code{dbMatrix} in order to be used for
densifying any \code{dbMatrix}. Should these params be less than the maximum
row and column indices the precomputed densification process will fail.

In such cases, run this function again with a larger
\code{n_rows} and \code{num_cols}, or to manually remove the precomputed
table set \code{options(dbMatrix.precomp = NULL)} in the R console.
}
\examples{
con = DBI::dbConnect(duckdb::duckdb(), ":memory:")
precompute(con = con , m = 100, n = 100, name = "precomputed_table")
}
\concept{dbMatrix}
\keyword{internal}
