% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_chunking.R
\name{preview_chunk_plan}
\alias{preview_chunk_plan}
\title{Plot a preview of the chunk plan}
\usage{
preview_chunk_plan(extent_list)
}
\arguments{
\item{extent_list}{list of extents from \code{chunk_plan}}
}
\description{
Plots the output from \code{\link{chunk_plan}} as a set of polygons to preview.
Can be useful for debugging. Invisibly returns the planned chunks as a SpatVector
of polygons
}
\keyword{internal}
