% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckDB_backend.R
\name{primaryKey}
\alias{primaryKey}
\title{Show if table has any primary keys}
\usage{
primaryKey(conn, remote_name)
}
\arguments{
\item{conn}{hashID of backend, DBI connection, or pool}

\item{remote_name}{name of table on DB}
}
\description{
Show if table has any primary keys
}
