% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_array}
\alias{print_array}
\title{Generate array for pretty printing of matrix values}
\usage{
print_array(
  i = NULL,
  j = NULL,
  x = NULL,
  dims,
  rownames = rep("", dims[1]),
  fill = ".",
  digits = 5L
)
}
\arguments{
\item{i, j, x}{matched vectors of integers in i and j, with value in x}

\item{dims}{dimensions of the array (integer vector of 2)}

\item{fill}{fill character}

\item{digits}{default = 5. If numeric, round to this number of digits}
}
\description{
Generate array for pretty printing of matrix values
}
\keyword{internal}
