% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query a database object}
\usage{
\S4method{query}{dbDataFrame,character}(x, statement, ...)
}
\arguments{
\item{x}{a database object}

\item{statement}{query statement in SQL}

\item{...}{additional params to pass}
}
\description{
Query a database object. Does not work in a lazy fashion.
}
