% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryStack-generic}
\alias{queryStack-generic}
\alias{queryStack,}
\alias{queryStack<-}
\alias{queryStack<-,dbData-method}
\title{Get and set the queryStack for a db backend object}
\usage{
\S4method{queryStack}{dbData}(x)

\S4method{queryStack}{dbData}(x) <- value
}
\arguments{
\item{x}{GiottoDB object}
}
\description{
Get and set the queryStack for a db backend object
}
