% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{raster_bin_data}
\alias{raster_bin_data}
\title{Bin database values in xy}
\usage{
raster_bin_data(data, extent, resolution = 50000)
}
\arguments{
\item{data}{tbl_Pool to use}

\item{extent}{terra extent in which to plot}

\item{resolution}{min number of tiles to collect}
}
\value{
matrix of rasterized point information
}
\description{
Bin database values in xy
}
\keyword{internal}
