% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{read_matrix}
\alias{read_matrix}
\title{read_matrix}
\usage{
read_matrix(con, value, name = "dbMatrix", overwrite = FALSE, ...)
}
\arguments{
\item{con}{DBI or duckdb connection object \code{(required)}}

\item{value}{path to .txt, .csv, .tsv or .gzip/.gz variants \code{(required)}}

\item{name}{name to assign file within database \code{(optional)}.
default: "dbMatrix"}

\item{overwrite}{whether to overwrite if `name` already exists in database.
\code{(required)}. default: FALSE}

\item{...}{additional params to pass}
}
\value{
tbl_dbi object
}
\description{
Read tabular matrix files into database
}
\details{
Construct a database VIEW of a .csv, .tsv, or .txt files or their .gz/.gzip
variants
}
\examples{
print('TODO')
}
