% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readMatrixDB}
\alias{readMatrixDB}
\title{Read matrix via R and data.table}
\usage{
readMatrixDB(
  path,
  backend_ID,
  remote_name = "temp",
  rowname_cols = 1,
  transpose = FALSE,
  ...
)
}
\arguments{
\item{path}{path to the matrix file}

\item{backend_ID}{ID of backend to use}

\item{remote_name}{name to assign the dbMatrix}

\item{rowname_cols}{numeric vector. Which columns of flat file contain rownames
information}

\item{transpose}{transpose matrix}

\item{...}{additional params to pass}
}
\value{
matrix
}
\description{
Function to read a matrix in from flat file directly into database.
This function is specific for tsv and csv style formats. The table will be read
in as 'read_temp' and converted to IJX format and permanently written in as a
table with the name supplied through remote_name param. The table generated
when reading ('read_temp') will then be dropped.
}
\details{
The matrix needs to have both unique column names and row names.
A Giotto database backend also must already be existing. .csv and .tsv files
when the database is duckdb will be handled through duckdb_read_csv
}
