% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readMatrixDT}
\alias{readMatrixDT}
\title{Read matrix via R and data.table}
\usage{
readMatrixDT(path, cores = 1L, transpose = FALSE)
}
\arguments{
\item{path}{path to the matrix file}

\item{cores}{number of cores to use}

\item{transpose}{transpose matrix}
}
\value{
matrix
}
\description{
Function to read a matrix in from flat file via R and data.table
}
\details{
The matrix needs to have both unique column names and row names
}
