% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{readMM}
\alias{readMM}
\title{read_MM}
\usage{
readMM(con, value, name = "dbMatrix", overwrite = FALSE, ...)
}
\arguments{
\item{con}{DBI or duckdb connection object \code{(required)}}

\item{value}{path to .mtx or .mtx.gz file \code{(required)}}

\item{name}{name to assign file within database \code{(optional)}.
default: "dbMatrix"}

\item{overwrite}{whether to overwrite if `name` already exists in database.
\code{(required)}. default: FALSE}

\item{...}{additional params to pass}
}
\value{
tbl_dbi object
}
\description{
Read matrix market file (.mtx or .mtx.gz)  into database
}
\details{
Construct a database VIEW of a .mtx or .mtx.gz file with columns 'i', 'j',
and 'x' representing the row index, column index, and value of the matrix,
respectively.

By default 'i' and 'j' are of type BIGINT and 'x' is of type DOUBLE.
Note: lack of support in R for BIGINT may cause errors when pulling data
into memory without proper type conversion.

By default, .mtx files are expected to contain two lines representing the
standard header information.
}
\examples{
print('TODO')
}
