% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{reconnectBackend-generic}
\alias{reconnectBackend-generic}
\alias{reconnectBackend}
\title{Reconnect GiottoDB backend}
\usage{
\S4method{reconnectBackend}{backendInfo}(x, with_login = FALSE, ...)
}
\arguments{
\item{x}{backendInfo object}

\item{with_login}{(default = FALSE) flag to check R environment variables
for login info and/or prompt for password}
}
\description{
Reconnect GiottoDB backend
}
