% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{remoteExistsTable}
\alias{remoteExistsTable}
\alias{remoteExistsTable,ANY,character-method}
\title{Whether a particular table exists in a connection}
\usage{
\S4method{remoteExistsTable}{dbData,character}(x, remote_name, ...)

\S4method{remoteExistsTable}{ANY,character}(x, remote_name, ...)
}
\arguments{
\item{x}{GiottoDB object or DB connection}

\item{...}{additional params to pass}
}
\description{
Whether a particular table exists in a connection
}
