% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{remoteListTables}
\alias{remoteListTables}
\alias{remoteListTables,ANY-method}
\title{List the tables in a connection}
\usage{
\S4method{remoteListTables}{dbData}(x, ...)

\S4method{remoteListTables}{ANY}(x, ...)
}
\arguments{
\item{x}{GiottoDB object or DB connection}

\item{...}{additional params to pass}
}
\description{
List the tables in a connection
}
