% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{remoteName-generic}
\alias{remoteName-generic}
\alias{remoteName}
\title{Database table name}
\usage{
\S4method{remoteName}{dbData}(x)
}
\arguments{
\item{x}{Duckling object}
}
\description{
Returns the table name within the database the the object acts as a link to
}
