% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{remoteValid}
\alias{remoteValid}
\title{Check if DB backend object has valid connection}
\usage{
\S4method{remoteValid}{dbData}(x)
}
\arguments{
\item{x}{GiottoDB object}
}
\description{
Check if DB backend object has valid connection
}
