% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{render_image}
\alias{render_image}
\title{internal function for rasterized rendering from matrix data}
\usage{
render_image(mat, cols = c("#002222", "white", "#800020"), axes = FALSE, ...)
}
\description{
internal function for rasterized rendering from matrix data
}
