% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{rowSums,dbDenseMatrix-method}
\alias{rowSums,dbDenseMatrix-method}
\alias{rowSums,dbSparseMatrix-method}
\alias{colSums,dbDenseMatrix-method}
\alias{colSums,dbSparseMatrix-method}
\alias{rowMeans,dbDenseMatrix-method}
\alias{rowMeans,dbSparseMatrix-method}
\alias{colMeans,dbDenseMatrix-method}
\alias{colMeans,dbSparseMatrix-method}
\title{Form Row and Column Sums and Means}
\usage{
\S4method{rowSums}{dbDenseMatrix}(x, na.rm = FALSE, dims = 1, ...)

\S4method{rowSums}{dbSparseMatrix}(x, na.rm = FALSE, dims = 1, ...)

\S4method{colSums}{dbDenseMatrix}(x, na.rm = FALSE, dims = 1, ...)

\S4method{colSums}{dbSparseMatrix}(x, na.rm = FALSE, dims = 1, ...)

\S4method{rowMeans}{dbDenseMatrix}(x, na.rm = FALSE, dims = 1, ...)

\S4method{rowMeans}{dbSparseMatrix}(x, na.rm = FALSE, dims = 1, ...)

\S4method{colMeans}{dbDenseMatrix}(x, na.rm = FALSE, dims = 1, ...)

\S4method{colMeans}{dbSparseMatrix}(x, na.rm = FALSE, dims = 1, ...)
}
\description{
See ?\link{\code{base::rowSums}} for more details.

See ?\link{\code{base::rowSums}} for more details.

See ?\link{\code{base::colSums}} for more details.

See ?\link{\code{base::colSums}} for more details.

See ?\link{\code{base::rowMeans}} for more details.

See ?\link{\code{base::rowMeans}} for more details.

See ?\link{\code{base::colMeans}} for more details.

See ?\link{\code{base::colMeans}} for more details.
}
\concept{summary}
