% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{save}
\alias{save}
\title{Compute or save a dbMatrix to disk}
\usage{
save(dbMatrix, name = "", overwrite = FALSE, ...)
}
\arguments{
\item{dbMatrix}{\code{dbMatrix} object}

\item{name}{name of table to save to disk}

\item{overwrite}{whether to overwrite if table already exists in database}

\item{...}{additional params to pass}

\item{description}{Saves dbMatrix table to specified name in the dbMatrix connection.
Note: Computing will not persist if the database is in ":memory:".}
}
\description{
Compute or save a dbMatrix to disk
}
\details{
This may break functions that rely on the existing table.
TODO: sync \code{dbMatrix} constructor with compute
}
\examples{
dbm <- sim_dbSparseMatrix()
compute(dbm, "new_table", overwrite = TRUE)
}
\keyword{internal}
