% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{colSds,dbDenseMatrix-method}
\alias{colSds,dbDenseMatrix-method}
\alias{colSds,dbSparseMatrix-method}
\alias{rowSds,dbDenseMatrix-method}
\alias{rowSds,dbSparseMatrix-method}
\title{Calculates the standard deviation for each row (column) of a matrix-like object}
\usage{
\S4method{colSds}{dbDenseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{colSds}{dbSparseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{rowSds}{dbDenseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{rowSds}{dbSparseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)
}
\description{
See ?\link{\code{MatrixGenerics::colSds}} for more details.

See ?\link{\code{MatrixGenerics::colSds}} for more details.

See ?\link{\code{MatrixGenerics::rowSds}} for more details.

See ?\link{\code{MatrixGenerics::rowSds}} for more details.
}
\concept{summary}
