% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatProxy.R
\name{simplify_poly_as_point}
\alias{simplify_poly_as_point}
\title{Represent polygon as single xy coordinate pair stand-ins}
\usage{
simplify_poly_as_point(
  x,
  method = c("mean", "lowL", "lowR", "upL", "upR"),
  ...
)
}
\arguments{
\item{x}{tbl_Pool (usually from a dbPolygonProxy object)}

\item{method}{one of 'mean', 'lowL', 'lowR', 'upL', 'upR'}

\item{...}{additional params to pass}
}
\value{
tbl_Pool of xy coordinates
}
\description{
Internal.
Calculate simple representative spatial locations for polygons.
Depending on \code{method} param, finds the polygon vertex that is the mean or
at the lower left, lower right, upper left, or upper right. All methods are
roughly equivalent in speed.
Note that this is NOT the polygon centroid.
}
\details{
Generate a representative point for all the polygons in a dbPolygonProxy. This
facilitates spatial selection of the polygons as there are fewer points to keep
track of, making it less computationally intensive and less ambiguous whether
a polygon should be selected by a specific selection if parts of it fall in multiple
selections.
}
\keyword{internal}
