% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_preset.R
\name{sql_create_table}
\alias{sql_create_table}
\alias{sql_create_pk_table}
\title{Generate sql to create a table with a primary key}
\usage{
sql_create_pk_table(
  con,
  table,
  fields,
  pk,
  row.names = NA,
  temporary = FALSE,
  ...
)
}
\arguments{
\item{con}{A database connection}

\item{table}{table name}

\item{fields}{a named character vector or a data frame}

\item{pk}{character. Which columns to select as the
primary key}

\item{row.names}{Either TRUE, FALSE, NA or a string.\cr
If TRUE, always translate row names to a column called "row_names". If FALSE,
never translate row names. If NA, translate rownames only if they're a
character vector. \cr
A string is equivalent to TRUE, but allows you to override the default name.\cr
For backward compatibility, NULL is equivalent to FALSE.}

\item{temporary}{If TRUE, will generate a temporary table statement.}

\item{...}{additional params to pass}
}
\description{
internal function taht allows the addition of a primary key line to the table
generation sql from sqlCreateTable
}
