% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sql_query-generic}
\alias{sql_query-generic}
\title{Query a database object}
\usage{
\S4method{sql_query}{dbData,character}(x, statement, ...)
}
\arguments{
\item{x}{a database object}

\item{statement}{query statement in SQL where the FROM field should be :data:}

\item{...}{additional params to pass}
}
\description{
Query a database object using a manually prepared SQL statement.
Integrates into dplyr chains. More general purpose than sql_gen* functions
}
\details{
Prepared SQL with properly quoted parameters are provided through the
\code{statement} param, where the FROM fields should use the token :data:.
This token is then replaced with the rendered SQL chain from the object being
queried, effectively appending the new instructions. This set of appended
instructions are then used to update the object.
}
\keyword{internal}
