% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{streamSpatialToDB_arrow}
\alias{streamSpatialToDB_arrow}
\title{Stream spatial data to database (arrow)}
\usage{
streamSpatialToDB_arrow(
  path,
  backend_ID,
  remote_name = "test",
  id_col = "poly_ID",
  xy_col = c("x", "y"),
  extent = NULL,
  file_format = NULL,
  chunk_size = 10000L,
  callback = NULL,
  overwrite = FALSE,
  custom_table_fields = NULL,
  custom_table_fields_attr = NULL,
  ...
)
}
\arguments{
\item{path}{path to database}

\item{backend_ID}{backend ID of database}

\item{remote_name}{name of table to generate in database backend}

\item{id_col}{ID column of data}

\item{xy_col}{character vector. Names of x and y value spatial coordinates or
vertices}

\item{extent}{terra SpatExtent (optional) that can be used to subset the data
to read in before it is saved to database}

\item{file_format}{(optional) file type (one of csv/tsv, arrow, or parquet)}

\item{chunk_size}{chunk size to use when reading in data}

\item{callback}{callback function to allow access to read-chunk-level data
formatting and filtering. Input and output should both be data.table}

\item{overwrite}{logical. whether to overwrite if table already exists in
database backend}

\item{custom_table_fields}{default = NULL. If desired, custom setup the fields
of the table. See \code{\link[DBI]{dbCreateTable}}}

\item{custom_table_fields_attr}{default = NULL. If desired, custom setup the
fields of the paired attributes table}
}
\value{
invisibly returns the final geom ID used
}
\description{
Stream spatial data to database (arrow)
}
