% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{sum,dbMatrix-method}
\alias{sum,dbMatrix-method}
\title{Sum of Vector Elements in dbMatrix objects}
\usage{
\S4method{sum}{dbMatrix}(x, na.rm = TRUE)
}
\arguments{
\item{x}{dbMatrix object}

\item{na.rm}{Always TRUE for dbMatrix queries. Included for compatibility
with the generic.}
}
\description{
\code{sum} returns the sum of all the values
  present in its arguments.
}
\concept{summary}
