% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatProxy.R
\name{svpoint_to_dt}
\alias{svpoint_to_dt}
\title{SpatVector point to data.table}
\usage{
svpoint_to_dt(spatvector, include_values = TRUE)
}
\arguments{
\item{spatvector}{spatvector to use}

\item{include_values}{whether to include attribute values in output}
}
\description{
SpatVector point to data.table
}
\keyword{internal}
