% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{t,dbMatrix-method}
\alias{t,dbMatrix-method}
\title{Matrix Transpose}
\usage{
\S4method{t}{dbMatrix}(x)
}
\arguments{
\item{x}{dbMatrix object}
}
\value{
dbMatrix object
}
\description{
Given a dbMatrix x, t returns the transpose of x.
}
\concept{transform}
