% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{tableBE}
\alias{tableBE}
\title{Create a table from database}
\usage{
tableBE(cPool, remote_name, ...)
}
\arguments{
\item{cPool}{database connection pool object}

\item{remote_name}{name of table with database}

\item{...}{additional params to pass to dplyr::tbl()}
}
\description{
Create a table from database
}
