% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckDB_backend.R
\name{tblName-generic}
\alias{tblName-generic}
\alias{tblName,dbMatrix-method}
\alias{tblName,dbDataFrame-method}
\title{Database table name}
\usage{
\S4method{tblName}{dbMatrix}(x)

\S4method{tblName}{dbDataFrame}(x)
}
\description{
Returns the table name within the database the the object acts as a link to
}
