% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{to_ijx_disk}
\alias{to_ijx_disk}
\title{to_ijx_disk}
\usage{
to_ijx_disk(con, name)
}
\arguments{
\item{con}{duckdb connection}

\item{name}{name of table to convert to ijx on disk}
}
\value{
remote table in long format unpivoted from wide format matrix
}
\description{
to_ijx_disk
}
\keyword{internal}
