% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{toDbDense}
\alias{toDbDense}
\title{Convert dbSparseMatrix to dbDenseMatrix}
\usage{
toDbDense(db_sparse)
}
\arguments{
\item{db_sparse}{dbSparseMatrix object.}
}
\description{
Convert a dbSparseMatrix to a dbDenseMatrix.
}
\examples{
TODO

}
