% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{validBE-generic}
\alias{validBE-generic}
\alias{validBE}
\title{Check if DB backend object has valid connection}
\usage{
\S4method{validBE}{dbData}(x)
}
\arguments{
\item{x}{Duckling object}
}
\description{
Check if DB backend object has valid connection
}
