% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{writeRemoteTable}
\alias{writeRemoteTable}
\alias{writeTableBE}
\title{Write a persistent table to the database}
\usage{
writeTableBE(cPool, remote_name, value, ...)
}
\arguments{
\item{cPool}{database connection pool object}

\item{remote_name}{name to assign the table within the database}

\item{value}{table to write}

\item{...}{additional params to pass to DBI::dbWriteTable()}
}
\description{
Write a persistent table to the database
}
