% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{closeBackend}
\alias{closeBackend}
\title{Close backend connection pool}
\usage{
closeBackend(backend_ID)
}
\arguments{
\item{backend_ID}{hashID of backend to close (optional)}
}
\description{
Closes pools. If specific backend_ID(s) are given then those will be closed.
When no specific ID is given, all existing backends will be closed.
}
