% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{cPool-generic}
\alias{cPool-generic}
\alias{cPool,}
\alias{cPool<-}
\alias{connection.tbl_Pool}
\alias{cPool<-,dbData-method}
\title{Database connection pool}
\usage{
\S4method{cPool}{dbData}(x)

connection.tbl_Pool(x)

\S4method{cPool}{dbData}(x) <- value
}
\arguments{
\item{x}{GiottoDB object}
}
\description{
Return the database connection pool object
}
