% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{createDBMatrix}
\alias{createDBMatrix}
\title{Create a matrix with database backend}
\usage{
createDBMatrix(
  matrix,
  remote_name = "test",
  db_path = ":temp:",
  dim_names,
  dims,
  cores = 1L,
  overwrite = FALSE,
  callback = callback_formatIJX(),
  ...
)
}
\arguments{
\item{matrix}{object coercible to matrix or filepath to matrix data accessible
by one of the read functions}

\item{remote_name}{name to assign within database}

\item{db_path}{path to database on disk}

\item{dim_names}{list of rownames and colnames of the matrix (optional)}

\item{dims}{dimensions of the matrix (optional)}

\item{cores}{number of cores to use if reading into R}

\item{overwrite}{whether to overwrite if table already exists in database}

\item{callback}{callback functions to apply to each data chunk before it is
sent to the database backend}

\item{...}{additional params to pass}
}
\description{
Create an S4 dbMatrix object that has a triplet format under the hood (ijx).
The data for the matrix is either written to a specified database file or
could also be read in from files on disk
}
