% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbDataFrame-class}
\alias{dbDataFrame-class}
\alias{dbDataFrame}
\title{S4 dbDataFrame class}
\description{
Representation of dataframes using an on-disk database. Each object
is used as a connection to a single table that exists within the database.
}
\section{Slots}{

\describe{
\item{\code{data}}{dplyr tbl that represents the database data}

\item{\code{hash}}{unique hash ID for backend}

\item{\code{remote_name}}{name of table within database that contains the data}

\item{\code{key}}{column to set as key for ordering and subsetting on i}
}}

