% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbPolygonProxy-class}
\alias{dbPolygonProxy-class}
\alias{dbPolygonProxy}
\title{S4 dbPolygonProxy class}
\description{
Representation of polygon information using an on-disk database. Intended to
be used to store information that can be pulled into terra polygon SpatVectors
}
\section{Slots}{

\describe{
\item{\code{data}}{lazy table containing geometry information with columns geom, part,
x, y, and hole}

\item{\code{attributes}}{dbDataFrame of attributes information, one of which (usually
the first) being 'ID' that can be joined/matched against the 'geom' values in
\code{attributes}}

\item{\code{n_poly}}{number of polygons}

\item{\code{poly_ID}}{polygon IDs}

\item{\code{extent}}{extent of polygons}

\item{\code{poly_filter}}{polygon SpatVector that is used to filter values on read-in}
}}

