% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{dim,dbMatrix-method}
\alias{dim,dbMatrix-method}
\title{Dimensions of an Object}
\usage{
\S4method{dim}{dbMatrix}(x)
}
\description{
See ?\link{\code{base::dim}} for more details.
}
\concept{matrix_props}
