% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{head,dbMatrix-method}
\alias{head,dbMatrix-method}
\alias{tail,dbMatrix-method}
\title{Return the First or Last Parts of an Object}
\usage{
\S4method{head}{dbMatrix}(x, n = 6L, ...)

\S4method{tail}{dbMatrix}(x, n = 6L, ...)
}
\description{
See ?\link{\code{utils::head}} for more details.

See ?\link{\code{utils::tail}} for more details.
}
\concept{matrix_props}
