% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{log,dbMatrix-method}
\alias{log,dbMatrix-method}
\title{Logarithms and Exponentials}
\usage{
\S4method{log}{dbMatrix}(x, ...)
}
\description{
See ?\link{\code{base::log}} for more details.
}
\concept{transform}
