% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{map_ijx_dimnames}
\alias{map_ijx_dimnames}
\title{Map dimnames to i,j indices}
\usage{
map_ijx_dimnames(dbMatrix, colName_i, colName_j)
}
\arguments{
\item{dbMatrix}{dbMatrix object}

\item{colName_i}{name of column rownames to add to database}

\item{colName_j}{name of column colnames to add to database
default: 'FALSE'.'}
}
\description{
Map dimnames to i,j indices
}
\details{
Constructs a table in a database that contains the accompanying dimnames
for a dbMatrix. The resulting columns in the table:
* i (row index)
* colName_i (rownames),
* j (col index)
* j_names (colnames)
* x (counts of i,j occcurences)
}
\keyword{internal}
