% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{mean,dbDenseMatrix-method}
\alias{mean,dbDenseMatrix-method}
\alias{mean,dbSparseMatrix-method}
\title{Arithmetic Mean}
\usage{
\S4method{mean}{dbDenseMatrix}(x, ...)

\S4method{mean}{dbSparseMatrix}(x, ...)
}
\description{
See ?\link{\code{base::mean}} for more details.

See ?\link{\code{base::mean}} for more details.
}
\concept{summary}
