% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{nrow}
\alias{nrow}
\alias{ncol}
\alias{nrow,dbDataFrame-method}
\alias{ncol,dbMatrix-method}
\title{The number of rows/cols}
\usage{
\S4method{nrow}{dbMatrix}(x)

\S4method{nrow}{dbDataFrame}(x)

\S4method{ncol}{dbMatrix}(x)
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns present in
\code{x}.
}
