% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{reconnect}
\alias{reconnect}
\title{Reconnect to database}
\usage{
\S4method{reconnect}{dbData}(x)
}
\arguments{
\item{x}{GiottoDB object}
}
\description{
Reconnects to a database. Additionally checks that the object contains a
reference to a valid table within the database. Runs at the beginning of
most function calls
}
