% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{colSds,dbDenseMatrix-method}
\alias{colSds,dbDenseMatrix-method}
\alias{colSds,dbSparseMatrix-method}
\alias{rowSds,dbDenseMatrix-method}
\alias{rowSds,dbSparseMatrix-method}
\title{Row (column) standard deviations for dbMatrix objects}
\usage{
\S4method{colSds}{dbDenseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{colSds}{dbSparseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{rowSds}{dbDenseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{rowSds}{dbSparseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{Always NULL for dbMatrix queries. Included for compatibility
with the generic.}

\item{na.rm}{Always TRUE for dbMatrix queries. Included for compatibility
with the generic.}

\item{center}{Always NULL for dbMatrix queries. Included for compatibility
with the generic.}

\item{...}{Additional arguments passed to specific methods.}

\item{useNames}{Always TRUE for dbMatrix queries. Included for compatibility
with the generic.}

\item{dim}{Always NULL for dbMatrix queries. Included for compatibility with
the generic.}
}
\description{
Calculates the standard deviation for each row (column) of a matrix-like
object.
}
\concept{summary}
