% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_objects}
\alias{simulate_objects}
\alias{simulate_duckdb}
\alias{simulate_dbDataFrame}
\alias{simulate_dbPointsProxy}
\alias{simulate_dbPolygonProxy}
\title{Simulate GiottoDB objects}
\usage{
simulate_duckdb(data = datasets::iris, name = "test", p = NULL)

simulate_dbDataFrame(data = NULL, name = "df_test", key = NA_character_)

simulate_dbPointsProxy(data = NULL)

simulate_dbPolygonProxy(data = NULL)
}
\arguments{
\item{data}{data to use}

\item{name}{remote name of table on database to set}
}
\description{
Create simulated GiottoDB objects in memory from pre-prepared data. Useful
for testing purposes and examples.
}
\section{Functions}{
\itemize{
\item \code{simulate_duckdb()}: Simulate a duckdb connection dplyr tbl_Pool in memory

\item \code{simulate_dbDataFrame()}: Simulate a dbDataFrame in memory

\item \code{simulate_dbPointsProxy()}: Simulate a dbPointsProxy in memory

\item \code{simulate_dbPolygonProxy()}: Simulate a dbPolygonProxy in memory

}}
