% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{tableInfo}
\alias{tableInfo}
\title{Get information about the table}
\usage{
tableInfo(conn, remote_name)
}
\arguments{
\item{conn}{hashID of backend, DBI connection or pool}

\item{remote_name}{name of table on DB}
}
\value{
a data.table of information about the existing DB tables
}
\description{
Get information about the table
}
