% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{tblName-generic}
\alias{tblName-generic}
\alias{tblName,}
\alias{tblName<-}
\alias{tblName<-,dbData-method}
\title{Database table name}
\usage{
\S4method{tblName}{dbData}(x)

\S4method{tblName}{dbData}(x) <- value
}
\arguments{
\item{x}{GiottoDB object}
}
\description{
Returns the table name within the database the the object acts as a link to
}
